started integrating the interpreter into the server. Needed to get signal/slot connectivity with the application.
