+++ instrumentserver/server/core.py
@@ -1,351 +1,91 @@
-import os
-import time
-import logging
+# -*- coding: utf-8 -*-
+"""
+Created on Mon Apr 20 15:57:23 2020
+
+@author: Chao
+"""
+import importlib
+import inspect
+import warnings
 import random
-from typing import Union, Optional
+import logging
+from types import MethodType
+from typing import Dict, Any, Union, Optional, Tuple, List
+from enum import Enum, unique
+from dataclasses import dataclass
+
+from jsonschema import validate as jsvalidate
 
 import zmq
-from qcodes import Parameter, Station, Instrument
+import jsonpickle
+import qcodes as qc
+from qcodes import (Station,
+                    Instrument,
+                    ChannelList,
+                    Parameter,
+                    ParameterWithSetpoints,
+                    Function)
+from qcodes.instrument import parameter
+from qcodes.utils.validators import Validator, Arrays
 
-from instrumentserver.server.interpreter import executeServerInstruction, instructionFromJson
-from instrumentserver import QtCore, QtWidgets, QtGui, serialize, resource
-from instrumentserver.base import send, recv
-from instrumentserver.log import LogLevels, LogWidget, log, setupLogging
-from instrumentserver.serialize import toParamDict
-from instrumentserver.helpers import toHtml
+from instrumentserver import serverInstructionSchema
+from .. import QtCore
+from ..log import LogLevels
+from ..base import send, recv
 
 
 logger = logging.getLogger(__name__)
 
+# for now, only these two types of parameter are supported, which should have
+# covered most of the cases)
+SUPPORTED_PARAMETER_CLASS = Union[Parameter, ParameterWithSetpoints]
 
-# TODO: parameter file location should be optionally configurable
-# TODO: add an option to save one file per station component
 
+# TODO: can we handle nested submodules right now?
 
-class StationList(QtWidgets.QTreeWidget):
-    """A widget that displays all objects in a qcodes station"""
+@unique
+class Operation(Enum):
+    """Valid operations for the server."""
 
-    cols = ['Name', 'Info']
+    #: get a list of instruments the server has instantiated
+    get_existing_instruments = 'get_existing_instruments'
 
-    #: Signal(str) --
-    #: emitted when a parameter or Instrument is selected
-    componentSelected = QtCore.Signal(str)
+    #: create a new instrument
+    create_instrument = 'create_instrument'
 
-    def __init__(self, parent=None):
-        super().__init__(parent)
+    #: get the blueprint of an instrument. The blueprint goes beyond a snapshot
+    #: and is useful for creating proxies.
+    get_instrument_blueprint = 'get_instrument_blueprint'
 
-        self.setColumnCount(len(self.cols))
-        self.setHeaderLabels(self.cols)
-        self.setSortingEnabled(True)
-        self.clear()
+    #: make a call to an object.
+    call = 'call'
 
-        self.itemSelectionChanged.connect(self._processSelection)
 
-    def clear(self):
-        super().clear()
-        self.instrumentsItem = QtWidgets.QTreeWidgetItem(['Instruments', ''])
-        self.paramsItem = QtWidgets.QTreeWidgetItem(['Parameters', ''])
-        self.addTopLevelItem(self.paramsItem)
-        self.addTopLevelItem(self.instrumentsItem)
+@dataclass
+class InstrumentCreationSpec:
+    instrument_class: str
+    args: Optional[Tuple] = None
+    kwargs: Optional[Dict[str, Any]] = None
 
-    def _addParameterTo(self, parent, obj):
-        lst = [obj.name]
-        info = toParamDict([obj], includeMeta=['vals', 'unit'])[obj.name]
-        infoString = f"{str(obj.__class__.__name__)}"
-        lst.append(infoString)
-        paramItem = QtWidgets.QTreeWidgetItem(lst)
-        parent.addChild(paramItem)
 
-    def addObject(self, obj: Union[Instrument, Parameter]):
-        lst = [obj.name, f"{str(obj.__class__.__name__)}"]
-        if isinstance(obj, Instrument):
-            insItem = QtWidgets.QTreeWidgetItem(lst)
-            self.instrumentsItem.addChild(insItem)
+@dataclass
+class ServerInstruction:
+    operation: Operation
+    create_instrument_spec: Optional[InstrumentCreationSpec] = None
+    call_spec: Optional[Any] = None
 
-        elif isinstance(obj, Parameter):
-            paramItem = QtWidgets.QTreeWidgetItem(lst)
-            self.paramsItem.addChild(paramItem)
 
-    def removeObject(self, name: str):
-        items = self.findItems(name, QtCore.Qt.MatchExactly | QtCore.Qt.MatchRecursive, 0)
-        for i in items:
-            if i not in [self.instrumentsItem, self.paramsItem]:
-                parent = i.parent()
-                parent.removeChild(i)
-                del i
+def instructionFromJson(d: Dict):
+    """Create an instruction object from a text-based json object.
 
-    def _processSelection(self):
-        items = self.selectedItems()
-        if len(items) == 0:
-            return
-        item = items[0]
-        if item not in [self.instrumentsItem, self.paramsItem]:
-            self.componentSelected.emit(item.text(0))
-
-
-class StationObjectInfo(QtWidgets.QTextEdit):
-
-    # TODO: make this a bit better looking
-    # TODO: add a TOC for the instrument?
-
-    @staticmethod
-    def htmlDoc(obj: Union[Parameter, Instrument]):
-        pass
-
-    @staticmethod
-    def parameterHtml(param: Parameter):
-        pass
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-
-        self.setReadOnly(True)
-
-    @QtCore.Slot(object)
-    def setObject(self, obj: Union[Instrument, Parameter]):
-        self.setHtml(toHtml(obj))
-
-
-class ServerStatus(QtWidgets.QWidget):
-    """A widget that shows the status of the instrument server."""
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-
-        self.layout = QtWidgets.QVBoxLayout(self)
-
-        # At the top: a status label, and a button for emitting a test message
-        self.addressLabel = QtWidgets.QLabel()
-        self.testButton = QtWidgets.QPushButton('Send test message')
-        self.statusLayout = QtWidgets.QHBoxLayout()
-        self.statusLayout.addWidget(self.addressLabel, 1)
-        self.statusLayout.addWidget(self.testButton, 0)
-        self.testButton.setSizePolicy(
-            QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed,
-                                  QtWidgets.QSizePolicy.Minimum)
-        )
-
-        self.layout.addLayout(self.statusLayout)
-
-        # next row: a window for displaying the incoming messages.
-        self.layout.addWidget(QtWidgets.QLabel('Messages:'))
-        self.messages = QtWidgets.QTextEdit()
-        self.messages.setReadOnly(True)
-        self.layout.addWidget(self.messages)
-
-    @QtCore.Slot(str)
-    def setListeningAddress(self, addr: str):
-        self.addressLabel.setText(f"Listening on: {addr}")
-
-    @QtCore.Slot(str, str)
-    def addMessageAndReply(self, message: str, reply: str):
-        tstr = time.strftime("%Y-%m-%d %H:%M:%S")
-        self.messages.setTextColor(QtGui.QColor('black'))
-        self.messages.append(f"[{tstr}]")
-        self.messages.setTextColor(QtGui.QColor('blue'))
-        self.messages.append(f"Server received: {message}")
-        self.messages.setTextColor(QtGui.QColor('green'))
-        self.messages.append(f"Server replied: {reply}")
-
-
-class ServerGui(QtWidgets.QMainWindow):
-    """Main window of the qcodes station server."""
-
-    serverPortSet = QtCore.Signal(int)
-
-    def __init__(self, station, startServer: Optional[bool] = True,
-                 serverPort: Optional[int] = 5555):
-        super().__init__()
-
-        self._paramValuesFile = os.path.join('.', 'parameters.json')
-        self._serverPort = serverPort
-
-        self.setWindowTitle('Instrument server')
-
-        # central widget is simply a tab container
-        self.tabs = QtWidgets.QTabWidget(self)
-        self.setCentralWidget(self.tabs)
-
-        self.stationList = StationList()
-        self.stationObjInfo = StationObjectInfo()
-        self.stationList.componentSelected.connect(self.displayComponentInfo)
-
-        stationWidgets = QtWidgets.QSplitter(QtCore.Qt.Horizontal)
-        stationWidgets.addWidget(self.stationList)
-        stationWidgets.addWidget(self.stationObjInfo)
-        stationWidgets.setSizes([300, 700])
-        self.tabs.addTab(stationWidgets, 'Station')
-
-        self.tabs.addTab(LogWidget(level=logging.INFO), 'Log')
-
-        self.serverStatus = ServerStatus()
-        self.tabs.addTab(self.serverStatus, 'Server')
-
-        # toolbar
-        self.toolBar = self.addToolBar('Tools')
-        self.toolBar.setIconSize(QtCore.QSize(16, 16))
-
-        # station tools
-        self.toolBar.addWidget(QtWidgets.QLabel('Station:'))
-        refreshStationAction = QtWidgets.QAction(
-            QtGui.QIcon(":/icons/refresh.svg"), 'Refresh', self)
-        refreshStationAction.triggered.connect(self.refreshStationComponents)
-        self.toolBar.addAction(refreshStationAction)
-
-        # parameter tools
-        self.toolBar.addSeparator()
-        self.toolBar.addWidget(QtWidgets.QLabel('Params:'))
-
-        loadParamsAction = QtWidgets.QAction(
-            QtGui.QIcon(":/icons/load.svg"), 'Load from file', self)
-        loadParamsAction.triggered.connect(self.loadParamsFromFile)
-        self.toolBar.addAction(loadParamsAction)
-
-        saveParamsAction = QtWidgets.QAction(
-            QtGui.QIcon(":/icons/save.svg"), 'Save to file', self)
-        saveParamsAction.triggered.connect(self.saveParamsToFile)
-        self.toolBar.addAction(saveParamsAction)
-
-        self.station = station
-        self.refreshStationComponents()
-
-        # A test client, just a simple helper object
-        self.testClient = TestClient()
-        self.testClient.log.connect(self.log)
-        self.serverPortSet.connect(self.testClient.setServerPort)
-        self.serverStatus.testButton.clicked.connect(
-            lambda x: self.testClient.sendMessage("Ping server.")
-        )
-
-        if startServer:
-            self.startServer()
-
-    def log(self, message, level=LogLevels.info):
-        log(logger, message, level)
-
-    def closeEvent(self, event):
-        if hasattr(self, 'stationServerThread'):
-            if self.stationServerThread.isRunning():
-                self.testClient.sendMessage(self.stationServer.SAFEWORD)
-        event.accept()
-
-    def startServer(self):
-        """Start the instrument server in a separate thread."""
-        self.stationServer = StationServer(station=self.station,
-                                           port=self._serverPort)
-        self.stationServerThread = QtCore.QThread()
-        self.stationServer.moveToThread(self.stationServerThread)
-        self.stationServerThread.started.connect(self.stationServer.startServer)
-        self.stationServer.finished.connect(lambda: self.log('ZMQ server closed.'))
-        self.stationServer.finished.connect(self.stationServerThread.quit)
-        self.stationServer.finished.connect(self.stationServer.deleteLater)
-
-        # connecting some additional things for messages
-        self.stationServer.log.connect(self.log)
-        self.stationServer.serverStarted.connect(self.serverStatus.setListeningAddress)
-        self.stationServer.serverStarted.connect(self._setServerAddr)
-        self.stationServer.finished.connect(
-            lambda: self.log('Server thread finished.', LogLevels.info)
-        )
-        self.stationServer.messageReceived.connect(self._messageReceived)
-        self.stationServerThread.start()
-
-    @QtCore.Slot(str)
-    def _setServerAddr(self, addr: str):
-        port = int(addr.split(":")[-1])
-        self._serverPort = port
-        self.serverPortSet.emit(port)
-
-    @QtCore.Slot(str, str)
-    def _messageReceived(self, message: str, reply: str):
-        maxLen = 80
-        messageSummary = message[:maxLen]
-        replySummary = reply[:maxLen]
-        self.log(f"Server received: {message}", LogLevels.debug)
-        self.log(f"Server replied: {reply}", LogLevels.debug)
-        self.serverStatus.addMessageAndReply(messageSummary, replySummary)
-
-    def addStationComponent(self, obj: Union[Parameter, Instrument]):
-        """Add an object (instrument or unbound parameter) to the station.
-
-        Will add the object also to the widget listing the station components.
-        """
-        self.station.add_component(obj)
-        self.log(f"Added to station: {obj}", LogLevels.info)
-        if isinstance(obj, Instrument) or isinstance(obj, Parameter):
-            self.stationList.addObject(obj)
-        else:
-            self.log(f"Cannot add <{obj}> to list of objects (unknown type).",
-                     LogLevels.warn)
-
-    def removeStationComponent(self, name: str):
-        """Remove an object from the station, and also from the widget showing
-        the station contents.
-
-        If the object is an instrument, we make sure to close it.
-        """
-        obj = self.station.components[name]
-        if isinstance(obj, Instrument):
-            self.station.close_and_remove_instrument(obj)
-        else:
-            self.station.remove_component(name)
-        self.stationList.removeObject(name)
-        self.log(f"Removed from station: {obj}", LogLevels.info)
-
-    def refreshStationComponents(self):
-        """clear and re-populate the widget holding the station components, using
-        the objects that are currently registered in the station."""
-        self.stationList.clear()
-        for _, obj in self.station.components.items():
-            self.stationList.addObject(obj)
-
-        for i in range(self.stationList.topLevelItemCount()):
-            item = self.stationList.topLevelItem(i)
-            self.stationList.expandItem(item)
-
-        for i, _ in enumerate(self.stationList.cols):
-            self.stationList.resizeColumnToContents(i)
-
-    def loadParamsFromFile(self):
-        """load the values of all parameters present in the server's params json file
-        to parameters registered in the station (incl those in instruments)."""
-
-        self.log(f"Loading parameters from file: "
-                 f"{os.path.abspath(self._paramValuesFile)}", LogLevels.info)
-        serialize.loadParamsFromFile(self._paramValuesFile, self.station)
-
-    def saveParamsToFile(self):
-        """save the values of all parameters registered in the station (incl
-         those in instruments) to the server's param json file."""
-
-        self.log(f"Saving parameters to file: "
-                 f"{os.path.abspath(self._paramValuesFile)}", LogLevels.info)
-        serialize.saveParamsToFile(self.station, self._paramValuesFile)
-
-    @QtCore.Slot(str)
-    def displayComponentInfo(self, name: Union[str, None]):
-        if name is not None:
-            obj = self.station.components[name]
-        else:
-            obj = None
-        self.stationObjInfo.setObject(obj)
-
-
-def startServer(station: Station, port: Optional[int] = 5555) -> "ServerGui":
-    """Create a server gui window
-
-    Can be used in an ipython kernel with Qt mainloop.
+    Translates string instructions to other types:
+        - operation is converted to :class:`.Operations`
     """
+    # TODO: validate json with schema?
 
-    setupLogging(addStreamHandler=False,
-                 logFile=os.path.abspath('instrumentserver.log'))
-    logging.getLogger('instrumentserver').setLevel(logging.DEBUG)
-    window = ServerGui(station, startServer=True, serverPort=port)
-    window.show()
-    return window
+    d['operation'] = Operation(d['operation'])
+    return ServerInstruction(**d)
 
-
-# initial prototype for server functionality below.
 
 class StationServer(QtCore.QObject):
     """Prototype for a server object.
@@ -373,6 +113,9 @@ class StationServer(QtCore.QObject):
 
     #: Signal() -- emitted when we shut down
     finished = QtCore.Signal()
+
+    #: Signal(Instrument) -- emitted when a new instrument was created
+    instrumentCreated = QtCore.Signal(Instrument)
 
     def __init__(self, station=None, parent=None, port=None):
         super().__init__(parent)
@@ -427,15 +170,13 @@ class StationServer(QtCore.QObject):
                 if message_ok:
                     # we don't need to use a try-block here, because
                     # errors are already handled in executeServerInstruction
-                    response_to_client = executeServerInstruction(
-                        self.station, instruction
-                    )
+                    response_to_client = self.executeServerInstruction(instruction)
                     response_log = f"Response to client: {str(response_to_client)}"
 
             else:
                 response_log = f"Invalid message type."
                 response_to_client = dict(error=response_log)
-            
+
             send(socket, response_to_client)
             self.messageReceived.emit(str(message), response_log)
 
@@ -443,30 +184,382 @@ class StationServer(QtCore.QObject):
         self.finished.emit()
         return True
 
+    def executeServerInstruction(self, instruction: ServerInstruction) \
+            -> Dict[str, Any]:
+        """
+        This is the interpreter function that the server will call to translate the
+        dictionary received from the proxy to instrument calls.
 
-class TestClient(QtCore.QObject):
-    """A simple client we can use to test if the server replies."""
+        :param station: qcodes.station object, the station that contains the
+            instrument to call.
+        :param instruction: The instruction object.
 
-    #: signal to emit log messages
-    log = QtCore.Signal(str, LogLevels)
-    serverAddressSet = QtCore.Signal()
+        :returns: the results returned from the instrument call
+        """
+        response = {'return_value': None,
+                    'error': None}
 
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.serverAddr = None
+        args = []
+        kwargs = {}
 
-    @QtCore.Slot(int)
-    def setServerPort(self, port: int):
-        self.serverAddr = f"tcp://localhost:{port}"
-        self.serverAddressSet.emit()
+        # we call a helper function depending on the operation that is requested
+        if instruction.operation == Operation.get_existing_instruments:
+            func = self._getExistingInstruments
+        elif instruction.operation == Operation.create_instrument:
+            func = self._createInstrument
+            args = [instruction.create_instrument_spec]
+        else:
+            raise NotImplementedError
 
-    @QtCore.Slot(str)
-    def sendMessage(self, msg: str):
-        context = zmq.Context()
-        socket = context.socket(zmq.REQ)
-        socket.connect(self.serverAddr)
-        self.log.emit(f"Test client sending request: {msg}", LogLevels.debug)
-        send(socket, msg)
-        reply = recv(socket)
-        self.log.emit(f"Test client received reply: {reply}", LogLevels.debug)
-        socket.close()
+        try:
+            returns = func(*args, **kwargs)
+            response['return_value'] = returns
+
+        except Exception as err:
+            response['error'] = err
+            warnings.simplefilter('always', UserWarning)  # since this runs in loop
+            warnings.warn(str(err))
+
+        return response
+
+    def _getExistingInstruments(self) -> Dict:
+        """
+        Get the existing instruments in the station,
+
+        :returns : a dictionary that contains the instrument name and its identity
+        """
+        instrument_info = {}
+        list_instruments = list(self.station.snapshot()['instruments'].keys())
+        for instrument_name in list_instruments:
+            instrument_info[instrument_name] = {}
+            instrument_info[instrument_name]['name'] = instrument_name
+
+            # TODO: Do we need the IDN?
+            instrument_info[instrument_name]['IDN'] = self.station[instrument_name].IDN()
+        return instrument_info
+
+    def _createInstrument(self, spec: InstrumentCreationSpec) -> None:
+        """Create a new instrument on the server"""
+
+        # TODO: Use the YAML configuration file for finding the correct package
+
+        sep_class = spec.instrument_class.split('.')
+        modName = '.'.join(sep_class[:-1])
+        clsName = sep_class[-1]
+        mod = importlib.import_module(modName)
+        cls = getattr(mod, clsName)
+
+        args = [] if spec.args is None else spec.args
+        kwargs = dict() if spec.kwargs is None else spec.kwargs
+
+        new_instrument = qc.find_or_create_instrument(
+            cls, *args, **kwargs)
+
+        self.instrumentCreated.emit(new_instrument)
+
+
+# ---------------- dictionary type definition -----------------------------------
+# TODO: migrate this to dataclasses?
+#   see: https://meeshkan.com/blog/typedict-vs-dataclasses-in-python/
+
+# class InstrumentCreateDictType(TypedDict):
+#     instrument_class: str
+#     args: Optional[tuple]
+#     kwargs: Optional[Dict]
+#
+#
+# class ParamDictType(TypedDict):
+#     name: str
+#     value: Optional[Any]
+#
+#
+# class FuncDictType(TypedDict):
+#     name: str
+#     args: Optional[tuple]
+#     kwargs: Optional[Dict]
+#
+#
+# class InstructionDictType(TypedDict):
+#     operation: Literal["get_existing_instruments",
+#                        'instrument_creation',  # TODO: rename
+#                        'proxy_construction',  # TODO: rename, clarify
+#                        'proxy_get_param',  # TODO: simplify -- only use callables
+#                        'proxy_set_param',  # TODO: simplify -- only use callables
+#                        'proxy_call_func',  # TODO: simplify -- only use callables
+#                        'proxy_write_raw',  #: TODO: needed?
+#                        'proxy_ask_raw',  # TODO: needed?
+#                        'instrument_snapshot']
+#
+#     instrument_name: Optional[str]  # not needed for 'get_existing_instruments'
+#     submodule_name: Optional[Union[str, None]]  # not needed for 'get_existing_instruments'
+#     instrument_create: Optional[InstrumentCreateDictType]  # for instrument creation
+#     parameter: Optional[ParamDictType]  # for get/set_param
+#     function: Optional[FuncDictType]  # for function call
+#     cmd: Optional[str]  # for function call
+
+
+
+
+# ------------------ helper functions -------------------------------------------
+def _processInstruction(station: Station,
+                        instructionDict: Dict):
+    """
+    process the operation instruction from the client and execute the operation.
+    
+    :param station: qcodes.station object, the station that contains the 
+        instrument to call.
+    :param instructionDict:  The dictionary passed from the instrument proxy
+        that contains the information needed for the operation. 
+
+    :returns: the results returned from the instrument call
+
+    """
+    operation = instructionDict['operation']
+    returns = None
+
+    if operation == 'get_existing_instruments':
+        # usually this operation will be called before all the others to check
+        # if the instrument already exists ont the server.
+        returns = _getExistingInstruments(station)
+
+    elif operation == 'instrument_creation':
+        _instrumentCreation(station, instructionDict)
+
+    else:  # doing operation on an existing instrument
+        instrument = station[instructionDict['instrument_name']]
+        if 'submodule_name' in instructionDict:
+            submodule_name = instructionDict['submodule_name']
+            if submodule_name is not None:
+                # do operation on an instrument submodule
+                instrument = getattr(instrument, submodule_name)
+
+        if operation == 'proxy_construction':
+            returns = _proxyConstruction(instrument)
+        elif operation == 'proxy_get_param':
+            returns = _proxyGetParam(instrument, instructionDict['parameter'])
+        elif operation == 'proxy_set_param':
+            _proxySetParam(instrument, instructionDict['parameter'])
+        elif operation == 'proxy_call_func':
+            returns = _proxyCallFunc(instrument, instructionDict['function'])
+        elif operation == 'proxy_write_raw':
+            returns = _proxyWriteRaw(instrument, instructionDict['cmd'])
+        elif operation == 'proxy_ask_raw':
+            returns = _proxyAskRaw(instrument, instructionDict['cmd'])
+        elif operation == 'instrument_snapshot':
+            returns = instrument.snapshot()
+        else:
+            # the instructionDict will be checked in the proxy before sent to
+            # here, so in principle this error should not happen.
+            raise TypeError('operation type not supported')
+    return returns
+
+
+
+
+
+def _proxyConstruction(instrument: Instrument) -> Dict:
+    """
+    Get the dictionary that describes the instrument.
+    This parameter part is similar to the qcodes snapshot of the instrument,
+    but with less information, also the string that describes the validator of
+    each parameter/argument is replaced with a jsonpickle encoded form so that
+    it is easier to reproduce the validator object in the proxy.
+    The functions are directly extracted from the instrument class.
+
+    :returns : a dictionary that describes the instrument
+
+    """
+    # get functions and parameters belong to the top instrument
+    construct_dict = _get_module_info(instrument)
+    # get functions and parameters belong to the submodule
+    try:
+        submodules = list(instrument.submodules.keys())
+    except AttributeError:
+        submodules = []
+
+    submodule_dict = {}
+    for module_name in submodules:
+        submodule = getattr(instrument, module_name)
+        # the ChannelList is not supported yet
+        if submodule.__class__ != ChannelList:
+            submodule_dict[module_name] = _get_module_info(submodule)
+
+    construct_dict['submodule_dict'] = submodule_dict
+    return construct_dict
+
+
+def _encodeArrayVals(param_vals: Arrays) -> Dict:
+    """ encode the array validators to a serializable format. This function is
+    necessary when the shape of the array validator contains a callable (another
+    parameter or a function) who belongs to an instrument class, which should
+    not be directly pickled.
+
+    :param param_vals: array validator of a parameter
+
+    :returns: A dictionary contains the encoded validator
+    """
+    encode_val = {"min_value": param_vals._min_value,
+                  "max_value": param_vals._max_value,
+                  "valid_types": param_vals.valid_types
+                  }
+    encode_val_shape = []
+    for dim in param_vals._shape:
+        # some possible ways of defining array validator shape in drivers.
+        # only parameters or functions of the same instrument are supported
+        if type(dim) is int:
+            encode_val_shape.append(dim)
+        elif isinstance(dim, Parameter) or isinstance(dim, parameter.GetLatest):
+            encode_val_shape.append(dim.name)
+        elif isinstance(dim, parameter._Cache):
+            encode_val_shape.append(dim._parameter.name)
+        elif type(dim) is MethodType:  # instrument function
+            encode_val_shape.append(dim.__name__)
+        else:
+            raise TypeError('Unsupported way of defining the shape of array '
+                            'validator, try to use one of the followings:\n'
+                            '1) a constant int\n'
+                            '2) a parameter in the same instrument\n'
+                            '   2a) self.parameter\n'
+                            '   2b) self.parameter.get_latest\n'
+                            '   2c) self.parameter.cache\n'
+                            '3) a function in the same instrument\n'
+                            )
+        encode_val['shape'] = encode_val_shape
+    return encode_val
+
+
+def _get_module_info(module: Instrument) -> Dict:
+    """ Get the parameters and functions of an instrument (sub)module and put
+    them a dictionary.
+    """
+    # get parameters
+    param_names = list(module.__dict__['parameters'].keys())
+    module_param_dict = {}
+    for param_name in param_names:
+        param: SUPPORTED_PARAMETER_CLASS = module[param_name]
+        param_dict_temp = {'name': param_name}
+        param_class = param.__class__
+        if param_class not in SUPPORTED_PARAMETER_CLASS.__args__:
+            raise TypeError(f"{param} is not supported yet, the current "
+                            f"supported parameter classes are "
+                            f"{SUPPORTED_PARAMETER_CLASS.__args__}")
+
+        if param_class is ParameterWithSetpoints:
+            param_dict_temp['setpoints'] = [setpoint.name for setpoint in
+                                            param.setpoints]
+
+        param_vals: Union[Validator, Arrays] = param.vals
+        try:  # directly pickle
+            param_dict_temp['vals'] = jsonpickle.encode(param_vals)
+        except RuntimeError:  # contains instrument class which cannot be pickled
+            if type(param_vals) is Arrays:
+                # Array validator is necessary for ParameterWithSetpoints
+                param_dict_temp['vals'] = _encodeArrayVals(param_vals)
+            else:  # otherwise, give up validation on proxy parameters
+                param_dict_temp['vals'] = jsonpickle.encode(None)
+
+        # some extra items are added here to support the snapshot in the proxy
+        # instrument class
+        param_dict_temp['class'] = param_class
+        param_dict_temp['unit'] = param.unit
+        param_dict_temp['snapshot_value'] = param._snapshot_value
+        param_dict_temp['snapshot_exclude'] = param.snapshot_exclude
+        param_dict_temp['max_val_age'] = param.cache._max_val_age
+        param_dict_temp['docstring'] = param.__doc__
+        module_param_dict[param_name] = param_dict_temp
+
+    # get functions
+    methods = set(dir(module))
+    base_methods = (dir(base) for base in module.__class__.__bases__)
+    unique_methods = methods.difference(*base_methods)
+    func_names = []
+    for method in unique_methods:
+        if callable(
+                getattr(module, method)) and method not in module_param_dict:
+            func_names.append(method)
+
+    module_func_dict = {}
+    for func_name in func_names:
+        func_dict_temp = {'name': func_name}
+        func = getattr(module, func_name)
+        func_dict_temp['docstring'] = func.__doc__
+
+        if func.__class__ == Function:
+            # for functions added using the old 'instrument.add_function'
+            # method, (the functions only have positional arguments, and each
+            # argument has a validator). In this case, the list of validators
+            # is pickled
+            func_dict_temp['arg_vals'] = jsonpickle.encode(func._args)
+        else:
+            # for functions added directly to instrument class as bound
+            # methods, the fullargspec and signature is stored in the
+            # dictionary(will be pickled when sending to proxy)(
+            # jsonpickle.en/decode has some bugs that don't worked for some
+            # function signatures)
+            jp_fullargspec = inspect.getfullargspec(func)
+            jp_signature = inspect.signature(func)
+            func_dict_temp['fullargspec'] = jp_fullargspec
+            func_dict_temp['signature'] = jp_signature
+        module_func_dict[func_name] = func_dict_temp
+    module_construct_dict = {'functions': module_func_dict,
+                             'parameters': module_param_dict}
+
+    return module_construct_dict
+
+#
+# def _proxyGetParam(instrument: Instrument, paramDict: ParamDictType) -> Any:
+#     """
+#     Get a parameter from the instrument.
+#
+#     :param paramDict: the dictionary that contains the parameter to change, the
+#         'value' item will be omitted.
+#     :returns : value of the parameter returned from instrument
+#
+#     """
+#     paramName = paramDict['name']
+#     return instrument[paramName]()
+#
+#
+# def _proxySetParam(instrument: Instrument, paramDict: ParamDictType) -> None:
+#     """
+#     Set a parameter in the instrument.
+#
+#     :param paramDict:  the dictionary that contains the parameter to change, the
+#         'value' item will be the set value.
+#         e.g. {'ch1': {'value' : 10} }
+#     """
+#     paramName = paramDict['name']
+#     instrument[paramName](paramDict['value'])
+#
+#
+# def _proxyCallFunc(instrument: Instrument, funcDict: FuncDictType) -> Any:
+#     """
+#     Call an instrument function.
+#
+#     :param funcDict:  the dictionary that contains the name of the function to
+#         call and the argument values
+#
+#     :returns : result returned from the instrument function
+#     """
+#     funcName = funcDict['name']
+#
+#     if ('kwargs' in funcDict) and ('args' in funcDict):
+#         args = funcDict['args']
+#         kwargs = funcDict['kwargs']
+#         return getattr(instrument, funcName)(*args, **kwargs)
+#     elif 'args' in funcDict:
+#         args = funcDict['args']
+#         return getattr(instrument, funcName)(*args)
+#     elif 'kwargs' in funcDict:
+#         kwargs = funcDict['kwargs']
+#         return getattr(instrument, funcName)(**kwargs)
+#     else:
+#         return getattr(instrument, funcName)()
+#
+#
+# def _proxyWriteRaw(instrument: Instrument, cmd: str) -> Union[str, None]:
+#     return instrument.write_raw(cmd)
+#
+#
+# def _proxyAskRaw(instrument: Instrument, cmd: str) -> Union[str, None]:
+#     return instrument.ask_raw(cmd)
